+++ stroom-core-client/src/main/java/stroom/query/client/presenter/DynamicFieldSelectionListModel.java
@@ -1,6 +1,6 @@
 package stroom.query.client.presenter;
 
-import stroom.datasource.api.v2.FindFieldInfoCriteria;
+import stroom.datasource.api.v2.FindFieldCriteria;
 import stroom.datasource.api.v2.QueryField;
 import stroom.docref.DocRef;
 import stroom.docref.StringMatch;
@@ -28,7 +28,8 @@ public class DynamicFieldSelectionListModel
     private final EventBus eventBus;
     private final DataSourceClient dataSourceClient;
     private DocRef dataSourceRef;
-    private FindFieldInfoCriteria lastCriteria;
+    private Boolean queryable;
+    private FindFieldCriteria lastCriteria;
     private final TaskListenerImpl taskListener = new TaskListenerImpl(this);
 
     @Inject
@@ -46,11 +47,12 @@ public class DynamicFieldSelectionListModel
                               final Consumer<ResultPage<FieldInfoSelectionItem>> consumer) {
         if (dataSourceRef != null) {
             final StringMatch stringMatch = StringMatch.contains(filter);
-            final FindFieldInfoCriteria findFieldInfoCriteria = new FindFieldInfoCriteria(
+            final FindFieldCriteria findFieldInfoCriteria = new FindFieldCriteria(
                     pageRequest,
                     null,
                     dataSourceRef,
-                    stringMatch);
+                    stringMatch,
+                    queryable);
 
             // Only fetch if the request has changed.
             if (!findFieldInfoCriteria.equals(lastCriteria)) {
@@ -84,6 +86,10 @@ public class DynamicFieldSelectionListModel
         this.dataSourceRef = dataSourceRef;
     }
 
+    public void setQueryable(final Boolean queryable) {
+        this.queryable = queryable;
+    }
+
     @Override
     public void reset() {
         lastCriteria = null;
@@ -91,7 +97,7 @@ public class DynamicFieldSelectionListModel
 
     @Override
     public void findFieldByName(final String fieldName, final Consumer<QueryField> consumer) {
-        dataSourceClient.findFieldByName(dataSourceRef, fieldName, consumer, taskListener);
+        dataSourceClient.findFieldByName(dataSourceRef, fieldName, queryable, consumer, taskListener);
     }
 
     @Override