+++ stroom-dashboard/stroom-storedquery-impl/src/main/java/stroom/storedquery/impl/StoredQueryModule.java
@@ -4,6 +4,7 @@ import stroom.job.api.ScheduledJobsBinder;
 import stroom.storedquery.api.StoredQueryService;
 import stroom.util.RunnableWrapper;
 import stroom.util.guice.RestResourcesBinder;
+import stroom.util.shared.scheduler.CronExpressions;
 
 import com.google.inject.AbstractModule;
 import jakarta.inject.Inject;
@@ -21,7 +22,7 @@ public class StoredQueryModule extends AbstractModule {
                 .bindJobTo(QueryHistoryClean.class, builder -> builder
                         .name("Query History Clean")
                         .description("Job to clean up old query history items")
-                        .cronSchedule("0 0 0 * * ?")
+                        .cronSchedule(CronExpressions.EVERY_DAY_AT_MIDNIGHT.getExpression())
                         .advanced(false));
     }