+++ stroom-index/stroom-index-impl/src/main/java/stroom/index/impl/IndexFieldDao.java
@@ -17,7 +17,7 @@
 package stroom.index.impl;
 
 
-import stroom.datasource.api.v2.FindIndexFieldCriteria;
+import stroom.datasource.api.v2.FindFieldCriteria;
 import stroom.datasource.api.v2.IndexField;
 import stroom.docref.DocRef;
 import stroom.util.shared.ResultPage;
@@ -28,5 +28,5 @@ public interface IndexFieldDao {
 
     void addFields(DocRef docRef, Collection<IndexField> fields);
 
-    ResultPage<IndexField> findFields(FindIndexFieldCriteria criteria);
+    ResultPage<IndexField> findFields(FindFieldCriteria criteria);
 }