+++ stroom-index/stroom-index-impl/src/main/java/stroom/index/impl/IndexFieldService.java
@@ -1,6 +1,6 @@
 package stroom.index.impl;
 
-import stroom.datasource.api.v2.FindIndexFieldCriteria;
+import stroom.datasource.api.v2.FindFieldCriteria;
 import stroom.datasource.api.v2.IndexField;
 import stroom.docref.DocRef;
 import stroom.query.common.v2.IndexFieldProvider;
@@ -12,7 +12,7 @@ public interface IndexFieldService extends IndexFieldProvider {
 
     void addFields(DocRef docRef, Collection<IndexField> fields);
 
-    ResultPage<IndexField> findFields(FindIndexFieldCriteria criteria);
+    ResultPage<IndexField> findFields(FindFieldCriteria criteria);
 
     void transferFieldsToDB(DocRef docRef);
 }