+++ stroom-index/stroom-index-lucene980/src/main/java/stroom/index/lucene980/Lucene980MemoryIndex.java
@@ -4,6 +4,7 @@ import stroom.datasource.api.v2.IndexField;
 import stroom.index.lucene980.SearchExpressionQueryBuilder.SearchExpressionQuery;
 import stroom.index.shared.LuceneIndexField;
 import stroom.query.api.v2.SearchRequest;
+import stroom.query.common.v2.IndexFieldCache;
 import stroom.search.extraction.FieldValue;
 import stroom.search.impl.SearchException;
 import stroom.util.logging.LambdaLogger;
@@ -17,7 +18,9 @@ import org.apache.lucene980.search.IndexSearcher;
 import org.apache.lucene980.search.TopDocs;
 
 import java.io.IOException;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 class Lucene980MemoryIndex implements stroom.search.extraction.MemoryIndex {
 
@@ -33,10 +36,13 @@ class Lucene980MemoryIndex implements stroom.search.extraction.MemoryIndex {
     @Override
     public boolean match(final SearchRequest searchRequest, final List<FieldValue> fieldValues) {
         final MemoryIndex memoryIndex = new MemoryIndex();
+        final Map<String, IndexField> indexFieldMap = new HashMap<>();
         for (final FieldValue fieldValue : fieldValues) {
             final IndexField indexField = fieldValue.field();
             final LuceneIndexField luceneIndexField = LuceneIndexField
                     .fromIndexField(indexField);
+
+            indexFieldMap.put(indexField.getFldName(), indexField);
 
             if (luceneIndexField.isIndexed()) {
                 final Analyzer fieldAnalyzer = searchExpressionQueryCache.getAnalyser(luceneIndexField);
@@ -48,12 +54,15 @@ class Lucene980MemoryIndex implements stroom.search.extraction.MemoryIndex {
         }
 
         // See if this set of fields matches the rule expression.
-        return matchQuery(searchRequest, memoryIndex);
+        final IndexFieldCache indexFieldCache = (key, fieldName) -> indexFieldMap.get(fieldName);
+        return matchQuery(searchRequest, memoryIndex, indexFieldCache);
     }
 
-    private boolean matchQuery(final SearchRequest searchRequest, final MemoryIndex memoryIndex) {
+    private boolean matchQuery(final SearchRequest searchRequest,
+                               final MemoryIndex memoryIndex,
+                               final IndexFieldCache indexFieldCache) {
         try {
-            final SearchExpressionQuery query = searchExpressionQueryCache.getQuery(searchRequest, true);
+            final SearchExpressionQuery query = searchExpressionQueryCache.getQuery(searchRequest, indexFieldCache);
             final IndexSearcher indexSearcher = memoryIndex.createSearcher();
             final TopDocs docs = indexSearcher.search(query.getQuery(), 100);