+++ stroom-index/stroom-index-mock/src/main/java/stroom/index/mock/MockIndexFieldService.java
@@ -17,7 +17,7 @@
 
 package stroom.index.mock;
 
-import stroom.datasource.api.v2.FindIndexFieldCriteria;
+import stroom.datasource.api.v2.FindFieldCriteria;
 import stroom.datasource.api.v2.IndexField;
 import stroom.docref.DocRef;
 import stroom.docref.StringMatch;
@@ -63,7 +63,7 @@ public class MockIndexFieldService implements IndexFieldService {
     }
 
     @Override
-    public ResultPage<IndexField> findFields(final FindIndexFieldCriteria criteria) {
+    public ResultPage<IndexField> findFields(final FindFieldCriteria criteria) {
         if (criteria.getDataSourceRef() != null && !loadedIndexes.contains(criteria.getDataSourceRef())) {
             transferFieldsToDB(criteria.getDataSourceRef());
             loadedIndexes.add(criteria.getDataSourceRef());
@@ -73,7 +73,12 @@ public class MockIndexFieldService implements IndexFieldService {
         final Set<IndexField> set = map.get(criteria.getDataSourceRef());
         final List<IndexField> filtered = set
                 .stream()
-                .filter(field -> stringMatcher.match(field.getFldName()).isPresent())
+                .filter(field -> {
+                    if (criteria.getQueryable() == null || criteria.getQueryable().equals(field.isIndexed())) {
+                        return stringMatcher.match(field.getFldName()).isPresent();
+                    }
+                    return false;
+                })
                 .toList();
         return ResultPage.createPageLimitedList(filtered, criteria.getPageRequest());
     }
@@ -98,11 +103,12 @@ public class MockIndexFieldService implements IndexFieldService {
 
     @Override
     public IndexField getIndexField(final DocRef docRef, final String fieldName) {
-        final FindIndexFieldCriteria findIndexFieldCriteria = new FindIndexFieldCriteria(
-                new PageRequest(0, 1),
+        final FindFieldCriteria findIndexFieldCriteria = new FindFieldCriteria(
+                PageRequest.oneRow(),
                 null,
                 docRef,
-                StringMatch.equals(fieldName));
+                StringMatch.equals(fieldName),
+                null);
         final ResultPage<IndexField> resultPage = findFields(findIndexFieldCriteria);
         if (resultPage.size() > 0) {
             return resultPage.getFirst();