+++ stroom-pipeline/src/main/java/stroom/pipeline/refdata/store/RefDataStoreModule.java
@@ -44,6 +44,7 @@ import stroom.pipeline.refdata.store.onheapstore.StringValueConsumer;
 import stroom.task.api.TaskTerminatedException;
 import stroom.util.RunnableWrapper;
 import stroom.util.guice.HasSystemInfoBinder;
+import stroom.util.shared.scheduler.CronExpressions;
 
 import com.google.inject.AbstractModule;
 import com.google.inject.assistedinject.FactoryModuleBuilder;
@@ -97,7 +98,7 @@ public class RefDataStoreModule extends AbstractModule {
                         .name(RefDataPurge.JOB_NAME)
                         .description("Purge old and partial reference data loads from the off heap store as " +
                                 "configured by 'purgeAge'.")
-                        .cronSchedule("0 0 2 * * ?"));
+                        .cronSchedule(CronExpressions.EVERY_DAY_AT_2AM.getExpression()));
     }