+++ stroom-state/stroom-state-impl/src/main/java/stroom/state/impl/StateSearchProvider.java
@@ -1,6 +1,6 @@
 package stroom.state.impl;
 
-import stroom.datasource.api.v2.FindFieldInfoCriteria;
+import stroom.datasource.api.v2.FindFieldCriteria;
 import stroom.datasource.api.v2.IndexField;
 import stroom.datasource.api.v2.QueryField;
 import stroom.docref.DocRef;
@@ -89,7 +89,7 @@ public class StateSearchProvider implements SearchProvider, IndexFieldProvider {
     }
 
     @Override
-    public ResultPage<QueryField> getFieldInfo(final FindFieldInfoCriteria criteria) {
+    public ResultPage<QueryField> getFieldInfo(final FindFieldCriteria criteria) {
         final StateDoc doc = getStateDoc(criteria.getDataSourceRef());
         final List<QueryField> fields = StateFieldUtil.getQueryableFields(doc.getStateType());
         return FieldInfoResultPageBuilder